+++ spec/integrations/rails_spec.rb
@@ -94,7 +94,7 @@ RSpec.describe "Rails integration" do
 
   subject { handler.handle(:command, request) }
 
-  before { allow(AnyCable.broadcast_adapter).to receive(:broadcast) }
+  before { allow(AnyCable).to receive(:broadcast) }
 
   let(:query) do
     <<~GQL
@@ -108,7 +108,7 @@ RSpec.describe "Rails integration" do
     GQL
   end
 
-  let(:redis) { AnycableSchema.subscriptions.redis }
+  let(:redis) { $redis }
 
   it "execute multiple clients + trigger + disconnect one by one" do
     # first, subscribe to obtain the connection state
@@ -129,7 +129,7 @@ RSpec.describe "Rails integration" do
     expect(redis.keys("graphql-subscriptions:*").size).to eq(1)
 
     schema.subscriptions.trigger(:post_updated, {id: "a"}, POSTS.first)
-    expect(AnyCable.broadcast_adapter).to have_received(:broadcast).once
+    expect(AnyCable).to have_received(:broadcast).once
 
     first_state = response.istate
 
@@ -144,7 +144,7 @@ RSpec.describe "Rails integration" do
     expect(redis.keys("graphql-subscriptions:*").size).to eq(1)
 
     schema.subscriptions.trigger(:post_updated, {id: "a"}, POSTS.first)
-    expect(AnyCable.broadcast_adapter).to have_received(:broadcast).twice
+    expect(AnyCable).to have_received(:broadcast).twice
 
     second_state = response_2.istate
 
@@ -164,6 +164,6 @@ RSpec.describe "Rails integration" do
     expect(redis.keys("graphql-subscriptions:*").size).to eq(0)
 
     schema.subscriptions.trigger(:post_updated, {id: "a"}, POSTS.first)
-    expect(AnyCable.broadcast_adapter).to have_received(:broadcast).twice
+    expect(AnyCable).to have_received(:broadcast).twice
   end
 end