+++ spec/redis_helper.rb
@@ -1,17 +1,15 @@
 # frozen_string_literal: true
 
-REDIS_TEST_DB_URL = "redis://localhost:6379/6"
+REDIS_TEST_DB_URL = ENV.fetch("REDIS_URL", "redis://localhost:6379/6")
 
-def setup_redis_test_db
-  test_url = ENV.fetch("REDIS_URL", REDIS_TEST_DB_URL)
-  channel = AnyCable.broadcast_adapter.channel
-  AnyCable.broadcast_adapter = :redis, {url: test_url, channel: channel}
-end
+channel = AnyCable.broadcast_adapter.channel
 
-setup_redis_test_db
+AnyCable.broadcast_adapter = :redis, {url: REDIS_TEST_DB_URL, channel: channel}
+
+$redis = Redis.new(url: REDIS_TEST_DB_URL)
 
 RSpec.configure do |config|
   config.before do
-    GraphQL::AnyCable.redis.flushdb
+    GraphQL::AnyCable.with_redis { _1.flushdb }
   end
 end